Add modal

+++ app/src/main/java/com/example/common_project01/ui/dbquery.kt
@@ -299,8 +299,6 @@ class DatabaseHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME
         val db = this.readableDatabase
         val cursor = db.rawQuery("SELECT * FROM DiaryData WHERE date = ? AND userId = ? ORDER BY id DESC", arrayOf(date,searchUserId)) // index 큰 순서로 정렬하여 최신 수정본 load
         var diaryData: DiaryData? = null
-        Log.d("myTag",searchUserId+"getget")
-
         if (cursor.moveToFirst()) {
             val userId = cursor.getString(cursor.getColumnIndex("userId"))
             val image = cursor.getString(cursor.getColumnIndex("image"))
@@ -313,9 +311,9 @@ class DatabaseHelper(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME
     }
     //.
 
-    fun deleteDiary(date: String) {
+    fun deleteDiary(date: String, userId: String) {
         val db = this.writableDatabase
-        db.delete("DiaryData", "date = ?", arrayOf(date))
+        db.delete("DiaryData", "date = ? AND userID = ?", arrayOf(date,userId))
         db.close()
     }
